Normalize wav reader, allow 8-bit and unsigned wav files.

+++ src/main/java/com/utsusynth/utsu/common/data/WavData.java
@@ -3,9 +3,9 @@ package com.utsusynth.utsu.common.data;
 /** Data for a single wav file. */
 public class WavData {
     private final double lengthMs;
-    private final int[] samples;
+    private final double[] samples;
 
-    public WavData(double lengthMs, int[] samples) {
+    public WavData(double lengthMs, double[] samples) {
         this.lengthMs = lengthMs;
         this.samples = samples;
     }
@@ -14,7 +14,7 @@ public class WavData {
         return lengthMs;
     }
 
-    public int[] getSamples() {
+    public double[] getSamples() {
         return samples;
     }
 }
